Add all delete logic (i think...)

+++ assets/js/message.js
@@ -3,6 +3,12 @@ import { getMessages } from "../app.js"
 $(document).ready(() => {
     const messages = document.querySelectorAll('.message-item')
 
+    const checkAll = document.querySelector('#check-all')
+    checkAll.addEventListener('click', () => selectAll())
+
+    const deleteAll = document.querySelector('#delete-all-messages')
+    deleteAll.addEventListener('click', () => deleteSelectedMessages())
+
     messages.forEach((item) => {
         const id = parseInt(item.id)
 
@@ -10,42 +16,60 @@ $(document).ready(() => {
         const messageCheck = item.querySelector('.message-check-container i')
         const dateElement = item.querySelector('small')
         const trashElement = item.querySelector('.fa-recycle')
-        const deleteAll = document.querySelector('#delete-all-messages')
-
-        deleteAll.addEventListener('click', () => deleteSelectedMessages())
 
         const messagesSelected = JSON.parse(localStorage.getItem('selected-messages')) ?? []
         const isSelected = messagesSelected.find((item) => item.id === id)
 
-        if (isSelected) changeSelected(card, messageCheck, dateElement, messagesSelected, deleteAll)
+        if (isSelected) changeSelected(card, messageCheck, dateElement, messagesSelected, deleteAll, checkAll)
 
         messageCheck.addEventListener('click', () => {
             const localMessages = JSON.parse(localStorage.getItem('selected-messages')) ?? []
 
             let element = localMessages.find((item) => item.id === id)
+            let newLocalMessages = element === undefined
+                ? (localMessages.push({ id }), localMessages)
+                : deleteLocalSelected(localMessages, element)
 
-            if (element === undefined) {
-                localMessages.push({ id })
-            } else {
-                const indexOf = localMessages.findIndex((item) => item === element)
+            newLocalMessages.length !== 0 ? localStorage.setItem('selected-messages', JSON.stringify(newLocalMessages)) : localStorage.clear()
 
-                localMessages.splice(indexOf, 1)
-            }
-
-            localMessages.length !== 0 ? localStorage.setItem('selected-messages', JSON.stringify(localMessages)) : localStorage.clear()
-
-            changeSelected(card, messageCheck, dateElement, localMessages, deleteAll)
+            changeSelected(card, messageCheck, dateElement, localMessages, deleteAll, checkAll)
         })
 
         trashElement.addEventListener('click', async () => {
             await deleteMessage(id)
+
+            const localMessages = JSON.parse(localStorage.getItem('selected-messages')) ?? []
+
+            let element = localMessages.find((item) => item.id === id)
+
+            if (element !== undefined) {
+                let newLocalMessages = deleteLocalSelected(localMessages, element)
+
+                const { length } = newLocalMessages
+
+                if (length !== 0) {
+                    localStorage.setItem('selected-messages', JSON.stringify(newLocalMessages))
+
+                } else {
+                    localStorage.clear()
+                    deleteAll.classList.add('disabled')
+                }
+            }
 
             getMessages()
         })
     })
 })
 
-const changeSelected = (card, messageCheck, dateElement, messages, deleteAll) => {
+const deleteLocalSelected = (localMessages, element) => {
+    const indexOf = localMessages.findIndex((item) => item === element)
+
+    localMessages.splice(indexOf, 1)
+
+    return localMessages
+}
+
+const changeSelected = (card, messageCheck, dateElement, messages, deleteAll, checkAll) => {
     card.classList.toggle('selected')
 
     messageCheck.classList.toggle('fa-square')
@@ -59,9 +83,37 @@ const changeSelected = (card, messageCheck, dateElement, messages, deleteAll) =>
 
     if (messages.length === 0) {
         deleteAll.classList.add('disabled')
+
+        checkAll.classList.add('d-none')
     } else {
-        if (deleteAll.classList.contains('disabled')) deleteAll.classList.remove('disabled')
+        if (deleteAll.classList.contains('disabled')) {
+            deleteAll.classList.remove('disabled')
+
+            checkAll.classList.remove('d-none')
+        }
     }
+}
+
+const selectAll = () => {
+    const checkAll = document.querySelector('#check-all')
+    const messages = document.querySelectorAll('.message-item')
+
+    checkAll.classList.toggle('fa-square-check')
+    checkAll.classList.toggle('fa-square')
+
+    checkAll.classList.toggle('fa-regular')
+    checkAll.classList.toggle('fa-solid')
+
+    messages.forEach((item) => {
+        const card = item.querySelector('.card')
+        const messageCheck = item.querySelector('.message-check-container i')
+
+        if (checkAll.classList.contains('fa-solid')) {
+            if (!card.classList.contains('selected')) messageCheck.click()
+        } else {
+            messageCheck.click()
+        }
+    })
 }
 
 const deleteMessage = async (id) => {
@@ -84,29 +136,39 @@ const deleteMessage = async (id) => {
 const deleteSelectedMessages = async () => {
     const selectedMessages = JSON.parse(localStorage.getItem('selected-messages'))
 
-    console.log(1)
+    /* const messages = selectedMessages.filter((item) => typeofdocument.getElementById(item.id) !== null) */
+
+    const isDelete = (item) => document.getElementById(item.id) !== null
+
+    const deletedMessages = selectedMessages.filter((item) => isDelete(item))
+    const unDeletedMessages = selectedMessages.filter((item) => !isDelete(item))
+
+    console.log(JSON.stringify(selectedMessages))
 
     await $.ajax({
         url: `/message/deleteSelectedMessages`,
         type: 'DELETE',
         contentType: 'application/json',
-        data: JSON.stringify(selectedMessages),
+        data: JSON.stringify(deletedMessages),
         success: (response) => {
-            console.log(response)
-           /*  if (response === '202') {
+            if (response === '202') {
                 const deleteAll = document.querySelector('#delete-all-messages')
+                const checkAll = document.querySelector('#check-all')
 
-                $(`.card.selected`).each((item) => item.remove())
+                $(`.message-item:has(.card.selected)`).each((index, item) => item.remove())
 
-                localStorage.clear()
+                localStorage.setItem('selected-messages', JSON.stringify(unDeletedMessages))
 
                 deleteAll.classList.add('disabled')
+                checkAll.classList.add('d-none')
+
+                getMessages()
             } else {
                 console.log('Error al borrar los mensajes :(')
-            } */
+            }
         },
         error: (err) => {
-            console.log('Error :(')
+            console.log(`Error :( ${err.responseText}`)
         }
     })
 }
